% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_beta.R
\name{compute_beta}
\alias{compute_beta}
\title{compute_beta}
\usage{
compute_beta(R, gam, dati)
}
\arguments{
\item{R:}{matrix of reference}

\item{gam:}{result of the svd}

\item{dati:}{array of the number}
}
\description{
Estimate beta for normalized analisys
}
