% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_norm.R
\name{compute_norm}
\alias{compute_norm}
\title{compute_norm}
\usage{
compute_norm(arraydata)
}
\arguments{
\item{arraydata:}{array of data}
}
\description{
Estimate the normalized mean of the array
}
