% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPA.R
\name{GPA}
\alias{GPA}
\title{Generilazed Procustes Analysis}
\usage{
GPA(arraydata, method = c("mean", "norm"), maxit = 10, mindist = 1,
  n = NULL)
}
\arguments{
\item{arraydata:}{array of data whit nrow,ncolum and the number of the matrix.}

\item{method:}{select a method for calculate GPA. "mean" used the matrix of mean for centrate the array of date.
"norm" used a normalized matrix of array, selected with the parameter n, for centrate the array of date.}

\item{maxit:}{maximum number of iteration inside the GPA funtion (default maxit=10).}

\item{mindist:}{select the minimum distance between the matrix (default mindist=1).}

\item{n:}{select the matrix to normalized for method "norm". It must be used into method "norm" (default n=NULL).}
}
\description{
The function that calculate the new ruotate arrays
}
