% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa_mean.R
\name{GPAmean}
\alias{GPAmean}
\title{GPAmean}
\usage{
GPAmean(arraydata, mindist, maxit)
}
\arguments{
\item{arraydata}{:initial array where i want to apply gpa}

\item{mindist:}{the minimal distace beetwen the matrix inside the array}

\item{maxit:}{maximum number of iteration}
}
\description{
The function that calculate the new ruotate arrays using the mean
}
