% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa_norm.R
\name{GPAnorm}
\alias{GPAnorm}
\title{GPAnorm}
\usage{
GPAnorm(arraydata, n, maxit)
}
\arguments{
\item{arraydata}{: array with the matrix of dates}

\item{n:}{selection of the matrix for calculate the norm}

\item{maxit:}{maximum number of the iteration}
}
\description{
The function that calculate the new ruotate arrays whit the normalizzed matrix
}
