% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procustemean.R
\name{procustemean}
\alias{procustemean}
\title{Rotazione Procuste tramite matrice media}
\usage{
procustemean(A, B)
}
\arguments{
\item{A}{array of matrix}

\item{B}{matrix of means}
}
\description{
The core of the GPAmean function
}
