% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procustenorm.R
\name{procustenorm}
\alias{procustenorm}
\title{Rotazione Procuste tramite matrice normalizzata}
\usage{
procustenorm(A, B)
}
\arguments{
\item{A}{array of matrix}

\item{B}{normalizzed matrix}
}
\description{
The core of the GPA function
}
