% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPA.R
\name{GPA}
\alias{GPA}
\title{Generilazed Procustes Analysis}
\usage{
GPA(arraydata, starting_point = c("mean", "matrix"), maxit = 10,
  mindist = 1/100, n = NULL, verbose = TRUE, normalization = FALSE)
}
\arguments{
\item{verbose}{TRUE}

\item{arraydata:}{array of data whit nrow,ncolum and the number of the matrix.}

\item{starting_point:}{select a starting_point for calculate GPA. "mean" used the matrix of mean for centrate the array of date.
"matrix" used a matrix of array, selected with the parameter n, for centrate the array of date.}

\item{maxit:}{maximum number of iteration inside the GPA funtion (default maxit=10).}

\item{mindist:}{select the minimum distance between the matrix (default mindist=1).}

\item{n:}{select the matrix for starting_point "matrix". It must be used into starting_point "norm" (default n=NULL).}

\item{normalization:}{Procuste distance with scaling (default FALSE)}
}
\description{
The function that calculate the new ruotate arrays
}
\examples{
arraydata=array(rnorm(10*50*4),c(10,50,4))
res=GPA(arraydata)
plot(res$epsilons,type="l")
}
